升级菜品数量可以为小数

+++ report/src/main/java/com/sygdsoft/controller/restaurant/SaleStreamReportController.java
@@ -70,7 +70,7 @@ public class SaleStreamReportController {
         List<FieldTemplate> templateList = new ArrayList<>();
         List<SaleStreamRow> saleStreamRowList = new ArrayList<>();
         Double totalConsume = 0.0;
-        Integer totalNum = 0;
+        Double totalNum = 0.0;
         String categoryParse = "";
         Double categoryConsume = 0.0;
         for (DeskDetailHistory deskDetailHistory : deskDetailHistoryList) {
