+++ sygd-core/src/main/java/com/sygdsoft/model/Menu.java
@@ -18,15 +18,15 @@ public class Menu extends BaseEntity {
     private Boolean foodSet;//是否是套餐
     private Boolean cargo;//库存货品
     private Double cost;//成本
-    private Integer remain;//剩余
+    private Double remain;//剩余
 
 
     public Menu() {
     }
 
-    public Integer getNotNullRemain(){
+    public Double getNotNullRemain(){
         if(remain==null){
-            return 0;
+            return 0.0;
         }else {
             return remain;
         }
@@ -136,11 +136,11 @@ public class Menu extends BaseEntity {
         this.cost = cost;
     }
 
-    public Integer getRemain() {
+    public Double getRemain() {
         return remain;
     }
 
-    public void setRemain(Integer remain) {
+    public void setRemain(Double remain) {
         this.remain = remain;
     }
 }
